Update scripts to support automated build
[git-p4: depot-paths = "//depot/ACG/Smith/Firmware/": change = 98721]


Former-commit-id: 5f3541c2d89f25bdc5dac27823fa432368af28a7

+++ deploy/build_package.rb
@@ -2,10 +2,37 @@
 
 require 'fileutils'
 require 'open-uri'
+require 'optparse'
+
+# Parse arguments
+options = {}
+
+OptionParser.new do |opts|
+
+  opts.on('--version VERSION', 'Specify version string to use in package name') do |version|
+    options[:version] = version
+  end
+
+  opts.on('--[no-]automate',
+          'Assume time is correct, remove existing file systems, and generate a new file system') do |automate|
+    options[:automate] = automate
+  end
+
+  opts.on('--smith-binary SMITH_BINARY',
+          'Path to smith binary to incorporate into build') do |smith_bin|
+    options[:smith_bin] = smith_bin
+  end
+
+  opts.on('--gem-cache-dir CACHE_DIR',
+          'Path to directory containing smith Ruby gem and dependencies to incorporate into build') do |gem_cache_dir|
+    options[:gem_cache_dir] = gem_cache_dir
+  end
+
+end.parse!
 
 # Read version from provided argument or from smith gem
-if ARGV[0]
-  version = ARGV.shift
+if options[:version]
+  version = options[:version]
 else
   require 'smith/version'
   version = Smith::VERSION
@@ -90,21 +117,35 @@ def run_command(cmd)
   print "\n"
 end
 
+def build_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file, redirect_output)
+  check_for_internet
+  
+  abort "#{oib_common_config_file} does not exist, aborting".red unless File.file?(oib_common_config_file)
+  abort "#{oib_config_file} does not exist, aborting".red unless File.file?(oib_config_file)
+  # Concatenate the common config options with the release specific options
+  File.write(oib_temp_config_file, "#{File.read(oib_common_config_file)}\n#{File.read(oib_config_file)}")
+
+  # Call to omap-image-builder
+  if redirect_output
+    puts "Executing omap-image-builder for #{oib_config_file}  See #{File.join(root, 'oib.log')} for output".green
+    oib_cmd = %Q(cd "#{root}" && omap-image-builder/RootStock-NG.sh -c #{oib_temp_config_file} > oib.log 2>&1)
+  else
+    puts "Executing omap-image-builder for #{oib_config_file}".green
+    # Redirect all output to stderr so output is printed to shell
+    # Otherwise Ruby will capture stdout
+    oib_cmd = %Q(cd "#{root}" && omap-image-builder/RootStock-NG.sh -c #{oib_temp_config_file} 1>&2)
+  end
+
+  %x(#{oib_cmd})
+  ensure_last_command_success(oib_cmd)
+  puts 'Operation complete'.green
+end
+
 def prompt_to_build_new_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file)
   print 'Do you want to build a new base filesystem with omap-image-builder? [y/N]: '.yellow
   if $stdin.gets.sub("\n", '').downcase == 'y'
     print "\n"
-    check_for_internet
-    # Call to omap-image-builder
-    puts "Executing omap-image-builder for #{oib_config_file}  See #{File.join(root, 'oib.log')} for output".green
-    abort "#{oib_common_config_file} does not exist, aborting".red unless File.file?(oib_common_config_file)
-    abort "#{oib_config_file} does not exist, aborting".red unless File.file?(oib_config_file)
-    # Concatenate the common config options with the release specific options
-    File.write(oib_temp_config_file, "#{File.read(oib_common_config_file)}\n#{File.read(oib_config_file)}")
-    oib_cmd = %Q(cd "#{root}" && omap-image-builder/RootStock-NG.sh -c #{oib_temp_config_file} > oib.log 2>&1)
-    %x(#{oib_cmd})
-    ensure_last_command_success(oib_cmd)
-    puts 'done'.green
+    build_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file, true)
   end
   print "\n"
 end
@@ -133,6 +174,16 @@ def list_filesystems(filesystems)
   print "\n"
 end
 
+def get_filesystems(deploy_dir)
+# Get a list of filesystems sorted from newest to oldest
+# Skip any non-directories
+Dir[File.join(deploy_dir, '/*')].
+  grep(/smith-release/).
+  reject { |d| !File.directory?(d) }.
+  sort_by { |d| File.stat(d).mtime }.
+  reverse
+end
+
 # Remove intermediate files if they exist any time the script exits
 at_exit do
   File.delete(oib_temp_config_file) if File.exist?(oib_temp_config_file)
@@ -150,22 +201,33 @@ end
 puts 'Smith firmware image builder script'
 print "\n"
 check_for_squashfs_tools
-check_date
-prompt_to_build_new_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file)
 
-# Get a list of filesystems sorted from newest to oldest
-# Skip any non-directories
-filesystems = Dir[File.join(deploy_dir, '/*')].
-  grep(/smith-release/).
-  reject { |d| !File.directory?(d) }.
-  sort_by { |d| File.stat(d).mtime }.
-  reverse.
-  map { |d| d.sub("#{deploy_dir}/", '') }
+if options[:automate]
+  puts "The system date is #{%x(date).sub("\n", '')}".green
+  print "\n"
+  
+  puts 'Removing all existing filesystems'.green
+  FileUtils.rm_r(get_filesystems(deploy_dir))
+  puts 'Operation complete'.green
+  print "\n"
+  
+  build_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file, false)
+else
+  check_date
+  prompt_to_build_new_filesystem(root, oib_config_file, oib_common_config_file, oib_temp_config_file)
+end
+
+filesystems = get_filesystems(deploy_dir).map { |d| d.sub("#{deploy_dir}/", '') }
 
 abort 'No base filesystem found, at least one filesystem must exist to continue, aborting'.red if filesystems.empty?
 
-list_filesystems(filesystems)
-selected_filesystem = filesystems[prompt_for_filesystem_index(filesystems)]
+if options[:automate]
+  selected_filesystem = filesystems.first
+else
+  list_filesystems(filesystems)
+  selected_filesystem = filesystems[prompt_for_filesystem_index(filesystems)]
+end
+
 selected_filesystem_root = Dir[File.join(deploy_dir, selected_filesystem, '/*')].grep(/rootfs/).first
 
 abort 'The selection does not contain a rootfs directory, aborting'.red if selected_filesystem_root.nil?
@@ -179,7 +241,8 @@ package_name = File.join(deploy_dir, "smith-#{version}.tar")
 
 puts 'Running install script'.green
 # Pass the install script the absolute path to the selected_filesystem_root
-run_command(%Q("#{File.join(script_dir, install_script_name)}" "#{selected_filesystem_root}"))
+# Also pass through optional paths to core components
+run_command(%Q("#{File.join(script_dir, install_script_name)}" "#{selected_filesystem_root}" "#{options[:smith_bin]}" "#{options[:gem_cache_dir]}"))
 
 puts "Building squashfs image (#{File.basename(image_temp_file)}) with #{selected_filesystem}".green
 # Remove any existing files; mksquashfs will attempt to append if the file exists
@@ -192,8 +255,10 @@ generate_md5sum_file(image_temp_file, md5sum_temp_file)
 puts 'Building package'.green
 run_command(%Q(cd "#{root}" && tar vcf "#{package_name}" "#{File.basename(image_temp_file)}" "#{File.basename(md5sum_temp_file)}"))
 
-puts 'Updating setup firmware'.green
-FileUtils.mkdir_p(firmware_setup_dir)
-run_command(%Q(rm -rfv "#{firmware_setup_dir}"/* && cp -v "#{image_temp_file}" "#{firmware_setup_dir}" && cp -v "#{md5sum_temp_file}" "#{firmware_setup_dir}/#{versions_file_name}"))
+if !options[:automate]
+  puts 'Updating setup firmware'.green
+  FileUtils.mkdir_p(firmware_setup_dir)
+  run_command(%Q(rm -rfv "#{firmware_setup_dir}"/* && cp -v "#{image_temp_file}" "#{firmware_setup_dir}" && cp -v "#{md5sum_temp_file}" "#{firmware_setup_dir}/#{versions_file_name}"))
+end
 
 puts "Successfully built #{package_name}, size: #{File.size(package_name) / 1048576}M".green
